`Period::contains` now accepts both `DateTimeInterface` and `Period`
