+++ tests/PrecisionTest.php
@@ -175,19 +175,6 @@ class PrecisionTest extends TestCase
     }
 
     /** @test */
-    public function precision_is_kept_when_testing_contains()
-    {
-        $a = Period::make('2018-01-01', '2018-01-31', Precision::DAY);
-
-        $this->assertTrue($a->contains(DateTimeImmutable::createFromFormat('Y-m-d H:i:s', '2018-01-01 00:00:00')));
-        $this->assertTrue($a->contains(DateTimeImmutable::createFromFormat('Y-m-d H:i:s', '2018-01-02 00:00:00')));
-        $this->assertTrue($a->contains(DateTimeImmutable::createFromFormat('Y-m-d H:i:s', '2018-01-31 23:59:59')));
-
-        $this->assertFalse($a->contains(DateTimeImmutable::createFromFormat('Y-m-d H:i:s', '2018-02-01 00:00:00')));
-        $this->assertFalse($a->contains(DateTimeImmutable::createFromFormat('Y-m-d H:i:s', '2017-12-21 23:59:59')));
-    }
-
-    /** @test */
     public function precision_is_kept_with_subtract()
     {
         $a = Period::make('2018-01-05 00:00:00', '2018-01-10 00:00:00', Precision::MINUTE);
